malloc and free finished

+++ TP2/Kernel/include/sbrk.h
@@ -2,6 +2,6 @@
 #define SBRK_H
 #include <stdint.h>
 
-void sbrk_handler(intptr_t increment, void ** buffer);
+void sbrk_handler(int increment, void ** buffer);
 
 #endif
